Correct errors in data and adds PPP conversion factors
I committed an error while creating data for RoW regions. Thus, I withdrew data for those variables that cannot be computed for RoW regions (because they are not extensive).
I added PPP conversion factors.
Note that with GDP, Population and PPP, the following variables can be deduced: GDP_PPP, GDPpc, GDPpcPPP. I let them in the data for the ease. Note that there is inconsistency in Cyprus data from the World Bank.