Corrected two country codes and added the link to download the data
Preliminary comparisons of GDP and Trade data computed from Exiobase (using my own original functions) gives me the impression that Exiobase data are not adjusted for be consistent with national accounts. 
Also, there seems to be a discrepancy between the values of imports and exports computed from Exiobase, and those from the World Bank, the former being lower of ~20%. I wonder from where it comes from. 
Finally, Exiobase data seems to be in basic price (as shows the fact that the data on value added corresponds to GDP at factor cost, not at GDP per se), and I wonder how is accounted the rest of GDP in the data (namely, value added taxes). 
Indeed, I am only one week new on pymrio so I have a lot to learn on all these things. In any case, I congratulate you, Konstantin and your colleagues, for this fantastic work!